data-pipeline/pyproject.toml
[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[project]
name = "pipeline"
version = "0.1.0"
description = "Murcko-scaffold aware assay processing pipeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "hydra-core>=1.3.2",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "polars>=0.20.0",
  "rdkit>=2023.9.5",
]

[project.scripts]
pipeline = "pipeline.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-cov>=4.1",
  "hypothesis>=6.98",
  "mypy>=1.9",
  "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S"]
ignore = ["E203"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
disallow_any_unimported = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_return_any = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_no_return = true
strict_equality = true
